% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dice_Roller.R
\name{roll_d20}
\alias{roll_d20}
\title{Roll d20}
\usage{
roll_d20(dice = 1, bonus = 0, each = FALSE, DC = NULL, AC = NULL,
  adv = NA)
}
\arguments{
\item{dice}{The number of dice to be rolled. Defaults to 1}

\item{bonus}{Any bonus amount to be added or subtracted from your roll. Defaults to 0}

\item{each}{Specifies if the bonus should be added to each individual die roll. If FALSE, the bonus will be added to the total rolls. Defaults to FALSE.}

\item{DC}{Allows user to enter a Difficulty Class to beat.}

\item{AC}{Allows user to enter an Armor Class to beat.}

\item{adv}{Used for advantage or disadvantage. If TRUE, dice will be rolled using advantage. If FALSE, dice will be rolled using disadvantage. Defaults to NA. This will only work if dice is set to 1.}
}
\description{
This function allows you to roll 20 sided dice.
}
\examples{
roll_d20(1, 5, adv=TRUE)
roll_d20()
}

