% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dice_Roller.R
\name{roll_dx}
\alias{roll_dx}
\title{Roll dx}
\usage{
roll_dx(x, dice = 1, bonus = 0, each = FALSE)
}
\arguments{
\item{x}{The number of sides on the dice.}

\item{dice}{The number of dice to be rolled. Defaults to 1}

\item{bonus}{Any bonus amount to be added or subtracted from your roll. Defaults to 0}

\item{each}{Specifies if the bonus should be added to each individual die roll. If FALSE, the bonus will be added to the total rolls. Defaults to FALSE.}
}
\description{
This function allows you to roll custom sided dice.
}
\examples{
roll_dx(3)
}

